chore: adjust -std

+++ src/2023/day-01/part-2/main.cpp
@@ -56,7 +56,7 @@ short find_number(string::iterator beginIterator, string::iterator endIterator, 
     if (is_number(*it))
       return to_number(*it);
 
-    for (short i = 0; i < WORDS.size(); i++)
+    for (size_t i = 0; i < WORDS.size(); i++)
     {
       string word = WORDS[i];
       if (has_word(word, it, endIterator, incrementBy))
